Fix implementation of weeks, months and years and round consistently
The implementation for longer periods introduced in smirik/php-datetime-ago#7
introduced a number of edge cases (including rendering periods of eg
1 month and 2 days as `2 days ago` due to the use of $diff->d rather than
$diff->days).

Additionally the calculation of periods did not match the documentation
and weeks, months and years were rounded up or down - this is inconsistent
with behaviour for smaller time periods.

Added additional tests for the polish translations to verify expected
results over a wider range of times and fixed the DateTimeAgo class
to pass all these specifications.