Support DateTimeInterface (requires PHP > 5.5)
Allow passing in either a DateTime or a DateTimeImmutable by
typehinting as the DateTimeInterface added in PHP5.5.

Bumps the bare minimum PHP version to 5.5.x (though all the 5.x
series are no longer supported anyway) and updates composer and
travis to officially test and support on 7.x.

This is backwards compatible for anyone simply using the class,
but is arguably a breaking change if anyone has extended the class
as the method signature of the ->get method has changed and will
produce an incompatible signatures warning...